# [code style] restruct fused_moe to avoid very long single file