--- a/python/sglang/srt/layers/moe/fused_moe_triton/__init__.py
+++ b/python/sglang/srt/layers/moe/fused_moe_triton/__init__.py
@@ -1,20 +1,14 @@
 from contextlib import contextmanager
 from typing import Any, Dict, Optional
 
-from .fused_moe import (
-    fused_experts,
-)
-from .fused_moe_triton_config import (
-    get_config_file_name,
-    try_get_optimal_moe_config,
-)
-from .moe_align_block_size import (
-    moe_align_block_size,
-)
 from sglang.srt.layers.moe.fused_moe_triton.layer import (
     FusedMoE,
     FusedMoeWeightScaleSupported,
 )
+
+from .fused_moe import fused_experts
+from .fused_moe_triton_config import get_config_file_name, try_get_optimal_moe_config
+from .moe_align_block_size import moe_align_block_size
 
 _config: Optional[Dict[str, Any]] = None