--- a/python/sglang/srt/layers/moe/fused_moe_triton/fused_moe.py
+++ b/python/sglang/srt/layers/moe/fused_moe_triton/fused_moe.py
@@ -13,17 +13,6 @@ import triton.language as tl
 
 from sglang.srt.layers.moe.moe_runner import MoeRunnerConfig
 from sglang.srt.layers.moe.topk import StandardTopKOutput
-from .fused_moe_triton_kernels import (
-    invoke_fused_moe_kernel,
-    moe_sum_reduce_triton,
-)
-from .moe_align_block_size import (
-    moe_align_block_size,
-)
-from .fused_moe_triton_config import (
-    try_get_optimal_moe_config,
-    get_config_dtype_str,
-)
 from sglang.srt.utils import (
     cpu_has_amx_support,
     direct_register_custom_op,
@@ -32,6 +21,10 @@ from sglang.srt.utils import (
     is_cuda,
     is_hip,
 )
+
+from .fused_moe_triton_config import get_config_dtype_str, try_get_optimal_moe_config
+from .fused_moe_triton_kernels import invoke_fused_moe_kernel, moe_sum_reduce_triton
+from .moe_align_block_size import moe_align_block_size
 
 _is_hip = is_hip()
 _is_cuda = is_cuda()
@@ -55,6 +48,7 @@ elif _is_hip:
         from vllm import _custom_ops as vllm_ops
 
 padding_size = 128 if bool(int(os.getenv("SGLANG_MOE_PADDING", "0"))) else 0
+
 
 def inplace_fused_experts(
     hidden_states: torch.Tensor,