<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

- unit-test

<img width="967" height="406" alt="图片" src="https://github.com/user-attachments/assets/2626a405-77a2-4faa-b21b-c5847f8a97f6" />

- MoE model acc test

```shell
python3 -m sglang.launch_server --model-path Qwen/Qwen2-57B-A14B-Instruct  --trust-remote-code  --tp 4
```

main:

```shell
➜  sglang git:(main) ✗  python3 benchmark/gsm8k/bench_sglang.py --num-questions 2000 --parallel 2000 --num-shots 8
100%|█████████████████████████████████████████████████████████████████████████████████████████████| 1319/1319 [00:18<00:00, 70.79it/s]
Accuracy: 0.701
Invalid: 0.020
Latency: 18.839 s
Output throughput: 9990.203 token/s
```

pr:

```shell
➜  sglang git:(main)  python3 benchmark/gsm8k/bench_sglang.py --num-questions 2000 --parallel 2000 --num-shots 8
100%|█████████████████████████████████████████████████████████████████████████████████████████████| 1319/1319 [00:19<00:00, 68.66it/s]
Accuracy: 0.709
Invalid: 0.026
Latency: 19.435 s
Output throughput: 9638.131 token/s
```

## Modifications

<!-- Detail the changes made in this pull request. -->

## Accuracy Tests

<!-- If this pull request affects model outputs (e.g., changes to the kernel or model forward code), provide accuracy test results. -->

## Benchmarking and Profiling

<!-- If this pull request impacts inference speed, provide benchmarking and profiling results. -->

## Checklist

- [ ] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [ ] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [ ] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [ ] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).